Add my version of CUDA Prim
Successfully tested and verified with graphs of up to 250K edges, but eg: 2.5M
edges segfaulted.

However, the segfault had nothing to do with CUDA. It was rather the
array creation in cudaRuntime() of evaluation.cu. Apparently, the arrays
got too large for the heap, so we might have to allocate them after all.

In any case, with only two levels of min-reduction, we can't do more
than an an MST with |V| = 1024*1024+1.