Don't test outside of the block (when blocksize is not power-2)
For example, when using a block of 6 elements ABCDEF with a block size
of 3,

Block 1  Block 2
-------  -------
A B C X  D E F Y
|/  |/   |/  |/
A   C    D   F
| /      | /
A        F
|    /
A

In block 1, comparing C to X, of F to Y, is incorrect.

Comparing C to X shouldn't be a big issue (because in contiguous memory, X=D)
and we should already be cutting of Y through the num_remaining test, but this
minor change resolved failing runs (where the determined minimum path is
a very large negative number).